+++ testproject/testproject/app/tests/test_models.py
@@ -1,9 +1,6 @@
 from datetime import datetime
-
-from django.test import TestCase
-from mock import patch
-
-from testproject.app.models import ExampleModel
+from django_cassandra_engine.test import TestCase
+from testproject.app.models import ExampleModel, ExampleModel2
 
 
 class ModelsTestCase(TestCase):
@@ -18,3 +15,20 @@ class ModelsTestCase(TestCase):
         obj = ExampleModel.objects.all()[0]
         self.assertEqual(obj.id, 1)
         self.assertEqual(obj.created_at, now)
+
+    def test_check_if_model_saved_to_test_keyspace(self):
+
+        now = datetime(2010, 1, 1, 1, 1)
+        obj_id = 123456
+        ExampleModel.objects.create(id=obj_id, created_at=now)
+
+        from cqlengine.connection import get_session
+        session = get_session()
+        session.set_keyspace('test_db')
+        self.assertEqual(
+            session.execute('SELECT id FROM example_model')[0]['id'], obj_id)
+
+    def test_truncate_models_before_running_tests_works(self):
+
+        self.assertEqual(ExampleModel.objects.count(), 0)
+        self.assertEqual(ExampleModel2.objects.count(), 0)